Whats new? 1-fixed bugs that caused the program to double write data to sheet when the button “enter” is used to execute the writing command 2-fixed bugs that caused program to open double windows of tableview when the shortcut command is used to open up the tableview window 3-changed the way delete method modifies the sheet and rows deleted. It used to delete rows completely and now it clears contents leaving the rows and their color there 4-New update was made that allows the program to be somewhat intuitive and realize that the name of the buyer exists in the sheet or doesn’t exist in the sheer and based on that it will either reset the slot#count or continue in case of a customer wanted to add more slots after an initial purchase was made

+++ src/main/java/org/example/Controller.java
@@ -115,7 +115,7 @@ public class Controller extends KrPreviewWindowController{
             if(event.getSource()==Write){
                 writeClicked();
             }
-            else if(event.getSource()==delete){
+            else if(event.getSource()==delete && delete.isFocused()){
                 deleteClicked();
             }
             else if(event.getSource()==update){
@@ -141,17 +141,6 @@ public class Controller extends KrPreviewWindowController{
                 clear4Clicked();
             }
         }
-        else if(event.isControlDown()){
-            if (event.getCode()==KeyCode.T){
-                previewClicked();
-            }
-        }
-        else if(event.isAltDown()){
-            if (event.getCode()==KeyCode.F4){
-                System.out.println("shortcut used to exit");
-                closeApplication();
-            }
-        }
     }
 
     @FXML
@@ -170,7 +159,6 @@ public class Controller extends KrPreviewWindowController{
             clearItemsInMethodChoiceBox(writeMethodChoiceBox);
             writeMethodChoiceBox.setTooltip(new Tooltip("Select a payment method"));
             addItemsToWritePaymentMethodChoiceBox(paymentMethodsArray);
-
             clearItemsInChoiceBox();
             sheetChoiceBox.setVisible(true);
             sheetLabel.setVisible(true);
